# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Practice generation, grading, progress tracking and study planning.",
    "title" : "IELTS Prep API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearer" : [ ]
  } ],
  "paths" : {
...
    "/api/reading/sessions/{id}" : {
      "get" : {
        "operationId" : "session_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReadingSessionView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Reading" ]
      }
    },
...
    "/api/writing/attempts/{id}/regrade" : {
      "post" : {
        "operationId" : "regrade",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WritingResultView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Writing" ]
      }
    },
...
    "/api/writing/sessions" : {
      "post" : {
        "operationId" : "start",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WritingStartRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WritingSessionView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Writing" ]
      }
    },
    "/api/writing/sessions/{id}" : {
      "get" : {
        "operationId" : "session",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WritingSessionView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Writing" ]
      }
    },
    "/api/writing/sessions/{id}/result" : {
      "get" : {
        "operationId" : "result",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WritingResultView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Writing" ]
      }
    },
    "/api/writing/sessions/{id}/submit" : {
      "post" : {
        "operationId" : "submit",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WritingSubmitRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WritingResultView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Writing" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ChartSeries" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "values" : {
            "type" : "array",
            "items" : {
              "type" : "number",
              "format" : "double"
            }
          }
        },
        "required" : [ "name", "values" ]
      },
      "CriterionBand" : {
        "type" : "object",
        "properties" : {
          "band" : {
            "type" : "integer",
            "format" : "int32"
          },
          "criterion" : {
            "type" : "string"
          },
          "justification" : {
            "type" : "string"
          },
          "strengths" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "weaknesses" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "band", "criterion", "justification", "strengths", "weaknesses" ]
      },
      "DiagramEdge" : {
        "type" : "object",
        "properties" : {
          "from" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "to" : {
            "type" : "string"
          }
        },
        "required" : [ "from", "label", "to" ]
      },
      "DiagramNode" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "x" : {
            "type" : "number",
            "format" : "double"
          },
          "y" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "required" : [ "id", "label", "x", "y" ]
      },
      "DiagramSpec" : {
        "type" : "object",
        "properties" : {
          "edges" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DiagramEdge"
            }
          },
          "nodes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DiagramNode"
            }
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "edges", "nodes", "title" ]
      },
...
      "MapFeature" : {
        "type" : "object",
        "properties" : {
          "h" : {
            "type" : "number",
            "format" : "double"
          },
          "kind" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "w" : {
            "type" : "number",
            "format" : "double"
          },
          "x" : {
            "type" : "number",
            "format" : "double"
          },
          "y" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "required" : [ "h", "kind", "label", "w", "x", "y" ]
      },
      "MapFigure" : {
        "type" : "object",
        "properties" : {
          "features" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MapFigureFeature"
            }
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "features", "title" ]
      },
      "MapFigureFeature" : {
        "type" : "object",
        "properties" : {
          "h" : {
            "type" : "number",
            "format" : "double"
          },
          "kind" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "w" : {
            "type" : "number",
            "format" : "double"
          },
          "x" : {
            "type" : "number",
            "format" : "double"
          },
          "y" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "required" : [ "h", "kind", "label", "w", "x", "y" ]
      },
      "MapMarker" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "x" : {
            "type" : "number",
            "format" : "double"
          },
          "y" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "required" : [ "key", "x", "y" ]
      },
      "MapSpec" : {
        "type" : "object",
        "properties" : {
          "features" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MapFeature"
            }
          },
          "markers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MapMarker"
            }
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "features", "markers", "title" ]
      },
...
      "Paragraph" : {
        "type" : "object",
        "properties" : {
          "label" : {
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          }
        },
        "required" : [ "label", "text" ]
      },
      "ProcessStep" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "step" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "description", "label", "step" ]
      },
      "Question" : {
        "type" : "object",
        "properties" : {
          "answers" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "justification_span" : {
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "options" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QuestionOption"
            }
          },
          "prompt" : {
            "type" : "string"
          }
        },
        "required" : [ "answers", "justification_span", "location", "number", "options", "prompt" ]
      },
      "QuestionGroup" : {
        "type" : "object",
        "properties" : {
          "choice" : {
            "type" : "boolean"
          },
          "context" : {
            "type" : "string"
          },
          "diagram" : {
            "$ref" : "#/components/schemas/DiagramSpec"
          },
          "flow_steps" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "group_id" : {
            "type" : "string"
          },
          "instructions" : {
            "type" : "string"
          },
          "map" : {
            "$ref" : "#/components/schemas/MapSpec"
          },
          "number_allowed" : {
            "type" : "boolean"
          },
          "options" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QuestionOption"
            }
          },
          "question_type" : {
            "type" : "string",
            "enum" : [ "MULTIPLE_CHOICE", "MULTIPLE_CHOICE_MULTI", "TRUE_FALSE_NOT_GIVEN", "YES_NO_NOT_GIVEN", "MATCHING_INFORMATION", "MATCHING_HEADINGS", "MATCHING_FEATURES", "MATCHING_SENTENCE_ENDINGS", "SENTENCE_COMPLETION", "SUMMARY_COMPLETION", "NOTE_COMPLETION", "TABLE_COMPLETION", "FLOW_CHART_COMPLETION", "DIAGRAM_LABEL_COMPLETION", "SHORT_ANSWER", "FORM_COMPLETION", "MAP_LABELLING", "MATCHING" ]
          },
          "questions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Question"
            }
          },
          "table" : {
            "$ref" : "#/components/schemas/TableSpec"
          },
          "word_limit" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "choice", "context", "diagram", "flow_steps", "group_id", "instructions", "map", "number_allowed", "options", "question_type", "questions", "table", "word_limit" ]
      },
      "QuestionOption" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          }
        },
        "required" : [ "key", "text" ]
      },
...
      "ReadingPassage" : {
        "type" : "object",
        "properties" : {
          "cefr" : {
            "type" : "string"
          },
          "difficulty" : {
            "type" : "integer",
            "format" : "int32"
          },
          "paragraphs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Paragraph"
            }
          },
          "question_groups" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/QuestionGroup"
            }
          },
          "title" : {
            "type" : "string"
          },
          "topic" : {
            "type" : "string"
          }
        },
        "required" : [ "cefr", "difficulty", "paragraphs", "question_groups", "title", "topic" ]
      },
...
      "ReadingPassageView" : {
        "type" : "object",
        "default" : null,
        "properties" : {
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "licence" : {
            "type" : [ "string", "null" ],
            "default" : ""
          },
          "numberOffset" : {
            "type" : "integer",
            "format" : "int32"
          },
          "passage" : {
            "$ref" : "#/components/schemas/ReadingPassage"
          },
          "questionCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sourceUrl" : {
            "type" : [ "string", "null" ],
            "default" : ""
          }
        },
        "required" : [ "itemId", "licence", "numberOffset", "passage", "questionCount", "sourceUrl" ]
      },
...
      "ReadingSessionView" : {
        "type" : "object",
        "default" : null,
        "properties" : {
          "kind" : {
            "type" : "string",
            "enum" : [ "READING_TEST", "READING_PASSAGE", "LISTENING_TEST", "LISTENING_SECTION", "WRITING_TASK1", "WRITING_TASK2", "WRITING_TEST", "SPEAKING_TEST", "SPEAKING_PART", "GRAMMAR_DIAGNOSTIC", "GRAMMAR_EXERCISE", "GRAMMAR_ERROR_DRILL", "VOCAB_REVIEW" ]
          },
          "mode" : {
            "type" : "string",
            "enum" : [ "PRACTICE", "EXAM" ]
          },
          "passages" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReadingPassageView"
            }
          },
          "sessionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string"
          },
          "timeLimitSeconds" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "default" : ""
          }
        },
        "required" : [ "kind", "mode", "passages", "sessionId", "startedAt", "status", "timeLimitSeconds" ]
      },
...
      "TableSpec" : {
        "type" : "object",
        "properties" : {
          "columns" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "rows" : {
            "type" : "array",
            "items" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "columns", "rows", "title" ]
      },
      "TaggedError" : {
        "type" : "object",
        "properties" : {
          "correction" : {
            "type" : "string"
          },
          "explanation" : {
            "type" : "string"
          },
          "original" : {
            "type" : "string"
          },
          "subtype" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "required" : [ "correction", "explanation", "original", "subtype", "type" ]
      },
      "Task1Academic" : {
        "type" : "object",
        "properties" : {
          "categories" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "chart_type" : {
            "type" : "string"
          },
          "figure_title" : {
            "type" : "string"
          },
          "key_features" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "maps" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MapFigure"
            }
          },
          "process_is_cycle" : {
            "type" : "boolean"
          },
          "process_steps" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProcessStep"
            }
          },
          "prompt" : {
            "type" : "string"
          },
          "series" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ChartSeries"
            }
          },
          "topic" : {
            "type" : "string"
          },
          "units" : {
            "type" : "string"
          },
          "x_label" : {
            "type" : "string"
          },
          "y_label" : {
            "type" : "string"
          }
        },
        "required" : [ "categories", "chart_type", "figure_title", "key_features", "maps", "process_is_cycle", "process_steps", "prompt", "series", "topic", "units", "x_label", "y_label" ]
      },
      "Task1General" : {
        "type" : "object",
        "properties" : {
          "bullet_points" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "key_features" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "letter_type" : {
            "type" : "string"
          },
          "prompt" : {
            "type" : "string"
          },
          "recipient" : {
            "type" : "string"
          },
          "situation" : {
            "type" : "string"
          },
          "topic" : {
            "type" : "string"
          }
        },
        "required" : [ "bullet_points", "key_features", "letter_type", "prompt", "recipient", "situation", "topic" ]
      },
      "Task2" : {
        "type" : "object",
        "properties" : {
          "essay_type" : {
            "type" : "string"
          },
          "key_features" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "prompt" : {
            "type" : "string"
          },
          "question" : {
            "type" : "string"
          },
          "statement" : {
            "type" : "string"
          },
          "subtopic" : {
            "type" : "string"
          },
          "topic" : {
            "type" : "string"
          }
        },
        "required" : [ "essay_type", "key_features", "prompt", "question", "statement", "subtopic", "topic" ]
      },
...
      "VocabUpgrade" : {
        "type" : "object",
        "properties" : {
          "better" : {
            "type" : "string"
          },
          "example" : {
            "type" : "string"
          },
          "original" : {
            "type" : "string"
          }
        },
        "required" : [ "better", "example", "original" ]
      },
      "WritingAttemptView" : {
        "type" : "object",
        "default" : null,
        "properties" : {
          "attemptId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "band" : {
            "type" : [ "number", "null" ],
            "format" : "double",
            "default" : ""
          },
          "criteriaBands" : {
            "type" : [ "object", "null" ],
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int32"
            },
            "default" : ""
          },
          "error" : {
            "type" : [ "string", "null" ],
            "default" : ""
          },
          "grade" : {
            "type" : "null",
            "$ref" : "#/components/schemas/WritingGrade",
            "default" : ""
          },
          "gradedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time",
            "default" : ""
          },
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "secondsSpent" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "default" : ""
          },
          "status" : {
            "type" : "string"
          },
          "submittedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "task" : {
            "type" : "integer",
            "format" : "int32"
          },
          "taskType" : {
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          },
          "wordCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "attemptId", "band", "criteriaBands", "error", "grade", "gradedAt", "itemId", "secondsSpent", "status", "submittedAt", "task", "taskType", "text", "wordCount" ]
      },
      "WritingGrade" : {
        "type" : "object",
        "properties" : {
          "criteria" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CriterionBand"
            }
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TaggedError"
            }
          },
          "improvements" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "model_answer" : {
            "type" : "string"
          },
          "overall_comment" : {
            "type" : "string"
          },
          "vocabulary_upgrades" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/VocabUpgrade"
            }
          },
          "word_count_comment" : {
            "type" : "string"
          }
        },
        "required" : [ "criteria", "errors", "improvements", "model_answer", "overall_comment", "vocabulary_upgrades", "word_count_comment" ]
      },
...
      "WritingResponseRequest" : {
        "type" : "object",
        "default" : null,
        "properties" : {
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "secondsSpent" : {
            "type" : "integer",
            "format" : "int32"
          },
          "text" : {
            "type" : "string"
          }
        }
      },
      "WritingResultView" : {
        "type" : "object",
        "default" : null,
        "properties" : {
          "attempts" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WritingAttemptView"
            }
          },
          "bandIsEstimate" : {
            "type" : "boolean"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "READING_TEST", "READING_PASSAGE", "LISTENING_TEST", "LISTENING_SECTION", "WRITING_TASK1", "WRITING_TASK2", "WRITING_TEST", "SPEAKING_TEST", "SPEAKING_PART", "GRAMMAR_DIAGNOSTIC", "GRAMMAR_EXERCISE", "GRAMMAR_ERROR_DRILL", "VOCAB_REVIEW" ]
          },
          "mode" : {
            "type" : "string",
            "enum" : [ "PRACTICE", "EXAM" ]
          },
          "sessionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "type" : "string"
          },
          "tasks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WritingTaskView"
            }
          },
          "timeLimitSeconds" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "default" : ""
          },
          "timeUsedSeconds" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "default" : ""
          },
          "writingBand" : {
            "type" : [ "number", "null" ],
            "format" : "double",
            "default" : ""
          }
        },
        "required" : [ "attempts", "bandIsEstimate", "kind", "mode", "sessionId", "status", "tasks", "timeLimitSeconds", "timeUsedSeconds", "writingBand" ]
      },
      "WritingSessionView" : {
        "type" : "object",
        "default" : null,
        "properties" : {
          "kind" : {
            "type" : "string",
            "enum" : [ "READING_TEST", "READING_PASSAGE", "LISTENING_TEST", "LISTENING_SECTION", "WRITING_TASK1", "WRITING_TASK2", "WRITING_TEST", "SPEAKING_TEST", "SPEAKING_PART", "GRAMMAR_DIAGNOSTIC", "GRAMMAR_EXERCISE", "GRAMMAR_ERROR_DRILL", "VOCAB_REVIEW" ]
          },
          "mode" : {
            "type" : "string",
            "enum" : [ "PRACTICE", "EXAM" ]
          },
          "sessionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string"
          },
          "tasks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WritingTaskView"
            }
          },
          "timeLimitSeconds" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "default" : ""
          }
        },
        "required" : [ "kind", "mode", "sessionId", "startedAt", "status", "tasks", "timeLimitSeconds" ]
      },
      "WritingStartRequest" : {
        "type" : "object",
        "default" : null,
        "properties" : {
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "mode" : {
            "type" : "string",
            "enum" : [ "PRACTICE", "EXAM" ]
          },
          "scope" : {
            "type" : "string",
            "enum" : [ "TASK1", "TASK2", "TEST" ],
            "default" : null
          }
        },
        "required" : [ "mode", "scope" ]
      },
      "WritingSubmitRequest" : {
        "type" : "object",
        "default" : null,
        "properties" : {
          "responses" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WritingResponseRequest"
            }
          },
          "timeUsedSeconds" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "WritingTaskView" : {
        "type" : "object",
        "default" : null,
        "properties" : {
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "minWords" : {
            "type" : "integer",
            "format" : "int32"
          },
          "minutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "task" : {
            "type" : "integer",
            "format" : "int32"
          },
          "task1Academic" : {
            "type" : "null",
            "$ref" : "#/components/schemas/Task1Academic",
            "default" : ""
          },
          "task1General" : {
            "type" : "null",
            "$ref" : "#/components/schemas/Task1General",
            "default" : ""
          },
          "task2" : {
            "type" : "null",
            "$ref" : "#/components/schemas/Task2",
            "default" : ""
          },
          "taskType" : {
            "type" : "string"
          }
        },
        "required" : [ "itemId", "minWords", "minutes", "task", "task1Academic", "task1General", "task2", "taskType" ]
      }
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}